requirements.txt
Flask
gradio_client
flask_cors
google-cloud-firestore
python-dotenv
firebase-admin
